# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pawnshop Management System API",
    "description": "A comprehensive API for managing pawnshop operations including loans, customers, items, payments, and sales.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@pawnshop.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8090",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type 'Bearer' followed by a space and the JWT token."
    }
  },
...
  "paths": {
...
    "/sales": {
      "get": {
        "tags": ["Sales"],
        "summary": "List Sales",
        "description": "Get a paginated list of sales",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of sales"
          }
        }
      },
      "post": {
        "tags": ["Sales"],
        "summary": "Create Sale",
        "description": "Create a new sale",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Sale created"
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "List Roles",
        "description": "Get a list of all roles",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of roles"
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Create Role",
        "description": "Create a new role",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Role created"
          }
        }
      }
    },
    "/roles/permissions": {
      "get": {
        "tags": ["Roles"],
        "summary": "List Available Permissions",
        "description": "Get a list of all available permissions",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of permissions"
          }
        }
      }
    },
...
  },
...
}